+++ src/Utils.h
@@ -89,8 +89,8 @@ public:
 float round(float number);
 #endif // _MSC_VER
 Point round(FPoint fp);
-FPoint screen_to_map(int x, int y, int camx, int camy);
-Point map_to_screen(float x, float y, int camx, int camy);
+FPoint screen_to_map(int x, int y, float camx, float camy);
+Point map_to_screen(float x, float y, float camx, float camy);
 Point center_tile(Point p);
 Point map_to_collision(FPoint p);
 FPoint collision_to_map(Point p);
